+++ test/integration/PaymentTypes/InvoiceSecuredTest.php
@@ -30,12 +30,18 @@ use UnzerSDK\Exceptions\UnzerApiException;
 use UnzerSDK\Resources\PaymentTypes\InvoiceSecured;
 use UnzerSDK\Resources\TransactionTypes\Charge;
 use UnzerSDK\test\BaseIntegrationTest;
+use UnzerSDK\test\Helper\TestEnvironmentService;
 
 /**
  * @deprecated since 1.2.0.0 PaylaterInvoice should be used instead in the future.
  */
 class InvoiceSecuredTest extends BaseIntegrationTest
 {
+    protected function setUp(): void
+    {
+        $this->getUnzerObject(TestEnvironmentService::getLegacyTestPrivateKey());
+    }
+
     /**
      * Verifies Invoice Secured payment type can be created.
      *