+++ test/integration/TransactionTypes/ShipmentTest.php
@@ -26,9 +26,15 @@ namespace UnzerSDK\test\integration\TransactionTypes;
 
 use UnzerSDK\Resources\PaymentTypes\InvoiceSecured;
 use UnzerSDK\test\BaseIntegrationTest;
+use UnzerSDK\test\Helper\TestEnvironmentService;
 
 class ShipmentTest extends BaseIntegrationTest
 {
+    protected function setUp(): void
+    {
+        $this->getUnzerObject(TestEnvironmentService::getLegacyTestPrivateKey());
+    }
+
     /**
      * Verify shipment transaction can be called.
      *